Move CLI flag validation to lib
