+++ packages/openui-cli/src/lib/resolve-args.ts
@@ -20,6 +20,16 @@ type ResolvedArgs<T extends Record<string, ArgDef<unknown>>> = {
   [K in keyof T]: T[K] extends { value: infer V } ? V : string;
 };
 
+export function rejectConflictingImmediateFlags(args: string[]): void {
+  const separatorIndex = args.indexOf("--");
+  const optionArgs = separatorIndex === -1 ? args : args.slice(0, separatorIndex);
+  const hasImmediate = optionArgs.some((arg) => arg === "--immediate" || arg === "-i");
+  const hasNoImmediate = optionArgs.includes("--no-immediate");
+  if (hasImmediate && hasNoImmediate) {
+    throw new CreateError("bad_args", "--immediate and --no-immediate cannot be used together.");
+  }
+}
+
 async function resolveOne(prompt: PromptConfig): Promise<string> {
   const { input, select } = await import("@inquirer/prompts");
   if (prompt.type === "select") {
